challenger: keep MultiField32 grind impl in grinding_challenger module
Move the GrindingChallenger impl for MultiField32Challenger back into
grinding_challenger.rs, next to the DuplexChallenger impl, so all
GrindingChallenger implementations stay grouped in one module
(addresses review feedback).

The optimized grind reaches private sponge fields only through the
pow_check_fn helper, which stays in multi_field_challenger.rs and is
now pub(crate). The trait impl itself touches no private state, so it
relocates cleanly while MultiField32Challenger keeps its fields private
(unlike DuplexChallenger, whose fields are pub).
